backend/requirements.txt
django
djangorestframework
django-cors-headers
groq
python-dotenv
requests
